+++ src/gql/types/PhoneNumberType.php
@@ -86,6 +86,10 @@ class PhoneNumberType
                     );
                 }
             ],
+            'timezones' => [
+                'name' => 'timezones',
+                'type' => Type::listOf(Type::string()),
+            ],
         ];
     }
 }