+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `getCarrierName`
 - Added `getExtension`
+- Added `getTimeZones`
 
 ## 2.0.0-beta.1 - 2022-03-22