+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `getExtension`
 - Added `getTimeZones`
 - Added `formatForCountry`
+- Added `formatForMobileDialing`
 
 ## 2.0.0-beta.1 - 2022-03-22